requirements/base.txt
Django==1.6.2
bpython==0.12
django-braces==1.2.2
django-model-utils==1.5.0
logutils==0.3.3
South==0.8.2

django-all-access==0.6.0
oauthlib==0.6.1
pycrypto==2.6.1
pytz==2014.1.1
requests==2.2.1
requests-oauthlib==0.4.0